fix: next-pwa 의존성 제거 및 커스텀 서비스워커 작성
