+++ next.config.ts
@@ -19,13 +19,15 @@ const nextConfig: NextConfig = {
   },
 }
 
-const withPWA = require('next-pwa')({
-  dest: 'public',
-  disable: process.env.NODE_ENV === 'development',
-  register: true,
-  scope: '/',
-  sw: 'sw.js',
-  customWorkerDir: 'src/worker',
-})
+// const withPWA = require('next-pwa')({
+//   dest: 'public',
+//   disable: process.env.NODE_ENV === 'development',
+//   register: true,
+//   scope: '/',
+//   sw: 'sw.js',
+//   customWorkerDir: 'src/worker',
+// })
 
-module.exports = withPWA(nextConfig)
+// module.exports = withPWA(nextConfig)
+
+export default nextConfig
