+++ package.json
@@ -14,13 +14,13 @@
     "autoprefixer": "^10.4.20",
     "classnames": "^2.5.1",
     "next": "^15.1.4",
-    "next-pwa": "^5.6.0",
     "nprogress": "^0.2.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.54.2",
     "react-toastify": "^11.0.5",
     "tailwindcss-preset-px-to-rem": "^1.2.1",
+    "web-push": "^3.6.7",
     "zustand": "^5.0.3"
   },
   "devDependencies": {
