+++ src/app/page.tsx
@@ -14,8 +14,9 @@ import { insertParticipant } from '@/lib/participant'
 import { useToast } from '@/components/common/toast/Toast'
 import {
   requestNotificationPermission,
-  sendPushNotification,
+  sendTestPushNotification,
 } from '@/utils/pushNotification'
+import { registerServiceWorker } from '@/utils/registerServiceWorker'
 
 const HomePage = () => {
   const user = useUserStore((state) => state.user)
@@ -43,8 +44,9 @@ const HomePage = () => {
     if (!user) return
     fetchNbreads(user.id)
 
+    registerServiceWorker()
     requestNotificationPermission()
-    sendPushNotification('테스트 알림', '테스트 알림입니다.')
+    sendTestPushNotification('테스트 알림', '테스트 알림입니다.')
   }, [user])
 
   // NbreadList가 업데이트된 후 totalAmount 계산
